1.PL Icon link  for windows set as comment. 2. fixed  dischargeDrone() function in BL_Actions, to calculate minutes and seconds as double type. 3. minor print fixes across a few toString() functions.

+++ BL/BL_Actions.cs
@@ -61,7 +61,7 @@ namespace BL
 
             IDAL.DO.DroneCharge tempCharge = myDal.GetDroneCharge(id);
             TimeSpan duration = DateTime.Now.Subtract((DateTime)tempCharge.EntranceTime);
-            double time = duration.Hours + (duration.Minutes / 60) + (duration.Seconds / 3600);
+            double time = duration.Hours + (double)duration.Minutes / 60 + (double)duration.Seconds / 3600;
             // update drone 
             drones[index].Battery = Math.Min(drones[index].Battery + (int)(droneHourlyChargeRate * time), 100);
             drones[index].Status = DroneStatus.Available;
